Rename to be less 6x-specific
