+++ imgcorrect/sensor_defs.py
