+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
-name = "correct6x"
+name = "imgcorrect"
 version = "1.3.0"
-description = "Library to perform various corrections on native 6X imagery"
+description = "Library to perform various corrections on imagery from supported sensors"
 authors = ["Your Name <you@example.com>"]
 
 [tool.poetry.dependencies]
