+++ scripts/correct_images.py
@@ -6,13 +6,13 @@ import sys
 from glob import glob
 
 import imgparse
-import correct6x
+import imgcorrect
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger()
 
 
-def correct_6x_images(input_path, calibration_id, output_path, no_ils_correct, no_reflectance_correct,
+def correct_images(input_path, calibration_id, output_path, no_ils_correct, no_reflectance_correct,
                       delete_original, exiftool_path, register):
 
     def _flag_format(flag):
@@ -37,48 +37,48 @@ def correct_6x_images(input_path, calibration_id, output_path, no_ils_correct, n
     logger.info("Delete original: %s", _flag_format(delete_original))
 
     # Read images:
-    image_df = correct6x.create_image_df(input_path, output_path)
+    image_df = imgcorrect.create_image_df(input_path, output_path)
 
     # Get image metadata:
     image_df['EXIF'] = image_df.image_path.apply(imgparse.get_exif_data)
 
     # Determine sensor type apply sensor specific settings
-    image_df = correct6x.apply_sensor_settings(image_df)
+    image_df = imgcorrect.apply_sensor_settings(image_df)
 
     # Get autoexposure correction:
     image_df['autoexposure'] = image_df.apply(lambda row: imgparse.get_autoexposure(row.image_path, row.EXIF), axis=1)
 
     # Split out calibration images, if present:
     if not no_reflectance_correct:
-        calibration_df, image_df = correct6x.create_cal_df(image_df, calibration_id)
+        calibration_df, image_df = imgcorrect.create_cal_df(image_df, calibration_id)
 
     # Get ILS correction:
     if not no_ils_correct:
-        image_df = correct6x.compute_ils_correction(image_df)
+        image_df = imgcorrect.compute_ils_correction(image_df)
     else:
         image_df['ILS_ratio'] = 1
 
     # Get reflectance correction:
     if not no_reflectance_correct:
-        image_df = correct6x.compute_reflectance_correction(image_df, calibration_df, not no_ils_correct)
+        image_df = imgcorrect.compute_reflectance_correction(image_df, calibration_df, not no_ils_correct)
     else:
         image_df['slope_coefficient'] = 1
 
     # Apply corrections:
-    image_df = image_df.apply(lambda row: correct6x.write_image(correct6x.apply_corrections(row), row), axis=1)
+    image_df = image_df.apply(lambda row: imgcorrect.write_image(imgcorrect.apply_corrections(row), row), axis=1)
 
     try:
         # Copy EXIF:
         logger.info("Writing EXIF data...")
-        image_df.apply(lambda row: correct6x.copy_exif(row, exiftool_path), axis=1)
+        image_df.apply(lambda row: imgcorrect.copy_exif(row, exiftool_path), axis=1)
 
         # Delete input imagery if requested:
         if delete_original:
-            correct6x.delete_all_originals(image_df)
+            imgcorrect.delete_all_originals(image_df)
 
         # Move output imagery to correct output directory:
         if (output_path and (output_path != input_path)) or delete_original:
-            correct6x.move_corrected_images(image_df)
+            imgcorrect.move_corrected_images(image_df)
 
         # Perform registration:
         if register:
@@ -101,7 +101,7 @@ if __name__ == '__main__':
     parser = argparse.ArgumentParser()
 
     parser.add_argument('input_path',
-                        help='Path to single-band, single-page .TIF files taken from a Sentera 6X sensor. Providing a '
+                        help='Path to image files taken from supported sensors. Providing a '
                              'file path to the original multi-page images is not currently supported. However, '
                              'specifying a folder containing all single-page files in their respective sub-folders '
                              'will cause the script to perform ILS correction recursively throughout each sub-folder.')
@@ -127,4 +127,4 @@ if __name__ == '__main__':
 
     args = parser.parse_args()
 
-    correct_6x_images(**vars(args))
+    correct_images(**vars(args))
